requirements.txt
requests>=2.25.0
python-dotenv>=0.15.0
tqdm>=4.50.0
